+++ CHANGELOG.md
@@ -292,3 +292,15 @@ async def post(myfile: UploadFile) -> Resp[Text, 200]:
     return "ok"
 ```
 
+
+## version 0.1.11
+
+### Improvements
+
+- add `middlewares: list[MiddlewareFactory[Any]] | None` param to the constructor of `Lihil` and `Route`, default to None
+
+- now `add_middleware` will append new middleware to the tail of call chain.
+
+- add `MiddlewareBuildError`, which will be raised when calling middleware factory fail 
+- add `NotSupportedError` for usage not currently supported, such a multiple return params.
+- add `InvalidParamType` for invalid param type, such as `Literal[3.14]`